inference: split runtime helpers out of training/
Carve a dep-light video_grouper/inference/ package (numpy + onnxruntime
+ cv2 only) and rewire the homegrown ball-tracking stages to use it via
static imports. This drops the importlib.import_module workaround in
detect.py / track.py and the --exclude-module fence in CI.

Why:
- A user who selects provider = homegrown from a bundled exe got
  ImportError because PyInstaller excluded training/torch/cv2/...
  from the modulegraph and the training/* package was lazy-loaded.
- training/ is for offline model development; runtime-only inference
  belongs in video_grouper/.

Changes:
- New: video_grouper/inference/{ball_detector,ball_tracker,field_detector}.py
  ported from training/inference/. ball_tracker.py replaces filterpy
  with a plain-numpy Kalman state. ultralytics CUDA-DLL side-effect
  dropped (torch supplies the same paths in the [ml] extra).
- Homegrown stages: static `from video_grouper.inference.* import ...`
  in detect.py and track.py.
- training/tasks/field_boundary.py imports from the new field_detector.
- training/cli/run_ball_detector.py is the new home for the
  `python -m ...` CLI; old training/inference/ entries are deleted.
- training/inference/panoramic_detector.py -> training/experiments/.
- training/inference/watchdog.py -> training/tools/onnx_watchdog.py.
- pyproject.toml: ultralytics moves out of main deps into [ml] only;
  numpy + opencv-python-headless added to main; scipy added to [ml].
- build-windows-service.yml: drop --exclude-module flags + workaround
  comment from both PyInstaller invocations.
- tests/: new inference tests for ball_detector / ball_tracker /
  field_detector. tests/conftest.py stubs onnxruntime to avoid a
  pytest-qt × onnxruntime-gpu DLL collision at collection time.

Verification:
- uv sync --extra dev --extra tray --extra service: torch + ultralytics
  NOT installed; cv2 + onnxruntime + numpy load OK.
- uv run pytest -m "not integration and not e2e": 856 passed,
  16 pre-existing failures (pydantic config, unrelated), no new breakage.
- ruff check / format: clean on touched files.
- git grep for `training.inference` / `from training` in video_grouper/:
  empty.
