+++ pyproject.toml
@@ -24,7 +24,8 @@ dependencies = [
     "webdriver-manager>=4.0.1",
     "pycryptodome>=3.20.0",
     "pillow>=11.3.0",
-    "ultralytics>=8.4.27",
+    "numpy>=2.2.0",
+    "opencv-python-headless>=4.13.0",
     "onnxruntime-gpu>=1.24.4; sys_platform == 'win32'",
     "uvicorn>=0.42.0",
     "fastapi>=0.135.2",
@@ -54,12 +55,13 @@ dev = [
 ml = [
     "torch>=2.7.0",
     "torchvision>=0.22.0",
-    "ultralytics>=8.3.70",
+    "ultralytics>=8.4.27",
     "opencv-python>=4.13.0",
     "numpy>=2.2.0",
     "supervision>=0.26.0",
     "lapx>=0.5.0",
     "filterpy>=1.4.5",
+    "scipy>=1.13",
     "fastapi>=0.115.0",
     "uvicorn>=0.34.0",
 ]
