+++ video_grouper/ball_tracking/providers/homegrown/stages/detect.py
@@ -1,12 +1,9 @@
 """Detection stage — run the homegrown ball detector on each frame.
 
-Wraps :mod:`training.inference.external_ball_detector`. Outputs
-per-frame detections to a ``detections.json`` next to the source.
-The result is a list of ``{frame_idx, cx, cy, w, h, conf}`` dicts in
-panoramic pixel coords.
-
-Heavy deps (``onnxruntime``, ``cv2``) are imported lazily inside the
-sync helper so the tray app doesn't load them unless this stage runs.
+Wraps :mod:`video_grouper.inference.ball_detector`. Outputs per-frame
+detections to a ``detections.json`` next to the source. The result is a
+list of ``{frame_idx, cx, cy, w, h, conf}`` dicts in panoramic pixel
+coords.
 """
 
 from __future__ import annotations
@@ -18,6 +15,7 @@ from pathlib import Path
 from typing import Any
 
 from video_grouper.ball_tracking.base import ProviderContext
+from video_grouper.inference.ball_detector import create_session, detect_video
 
 from . import register_stage
 from .base import ProcessingStage
@@ -33,21 +31,10 @@ def _run_detection(
     frame_interval: int,
     use_gpu: bool,
 ) -> int:
-    """Sync helper: run detection, write JSON. Returns detection count.
-
-    Uses ``importlib.import_module`` rather than a ``from … import`` to keep
-    the heavy training-pipeline deps (torch, ultralytics, opencv) outside
-    PyInstaller's static modulegraph. Otherwise the service / tray exes
-    balloon past NSIS's 32-bit mmap ceiling at install time.
-    """
-    import importlib
-    from pathlib import Path as _Path
-
-    ext_ball = importlib.import_module("training.inference.external_ball_detector")
-
-    sess = ext_ball.create_session(_Path(model_path), use_gpu=use_gpu)
-    detections = ext_ball.detect_video(
-        _Path(video_path),
+    """Sync helper: run detection, write JSON. Returns detection count."""
+    sess = create_session(Path(model_path), use_gpu=use_gpu)
+    detections = detect_video(
+        Path(video_path),
         sess,
         frame_interval=frame_interval,
         conf_threshold=confidence,
